Lab4/ark-pzpi-23-5-melnyk-anastasiia-lab4/emulator/requirements.txt
paho-mqtt==1.6.1
getmac==0.9.4
python-dotenv==1.0.0
requests==2.31.0